[clang][Modules] Move `ASTSourceDescriptor` into its own file
